separated evaluator classes into sub files
