+++ benchllm/evaluator/__init__.py
@@ -1,0 +1,3 @@
+from benchllm.evaluator.evaluator import Evaluator  # noqa
+from benchllm.evaluator.semantic import SemanticEvaluator  # noqa
+from benchllm.evaluator.string_match import StringMatchEvaluator  # noqa
