+++ benchllm/evaluator/evaluator.py
@@ -5,7 +5,7 @@ from itertools import groupby
 from operator import attrgetter
 from pathlib import Path
 from timeit import default_timer as timer
-from typing import List, Optional
+from typing import Optional
 
 import yaml
 from pydantic import BaseModel
@@ -13,7 +13,6 @@ from pydantic import BaseModel
 from benchllm.data_types import Evaluation, FunctionID, Prediction
 from benchllm.input_types import Json
 from benchllm.listener import EvaluatorListener
-from benchllm.similarity import semantically_similar
 
 
 class Evaluator(ABC):
@@ -83,6 +82,10 @@ class Evaluator(ABC):
     def workers(self) -> int:
         return self._workers
 
+    @property
+    def predictions(self) -> list[Prediction]:
+        return self._predictions
+
     @abstractmethod
     def evaluate_prediction(self, prediction: Prediction) -> Optional[Match]:
         """Evaluate a single prediction, return a Match if the prediction matches the expected output."""
@@ -114,62 +117,3 @@ class Evaluator(ABC):
     def _broadcast_evaluate_ended(self, evaluations: list[Evaluation]) -> None:
         for listener in self._listeners:
             listener.evaluate_ended(evaluations)
-
-
-class SemanticEvaluator(Evaluator):
-    def __init__(self, *, model: str = "gpt-3", workers: int = 1):
-        super().__init__(workers=workers)
-        self.model = model
-
-    def evaluate_prediction(self, prediction: Prediction) -> Optional[Evaluator.Match]:
-        for expected in prediction.test.expected:
-            if semantically_similar(expected, prediction.output, model=self.model):
-                return Evaluator.Match(prediction=prediction.output, expected=expected)
-        return None
-
-
-class StringMatchEvaluator(Evaluator):
-    def __init__(self, *, case_sensitive: bool = False, fuzzy: bool = False, workers: int = 1):
-        super().__init__(workers=workers)
-
-        self._case_sensitive = case_sensitive
-        self._fuzzy = fuzzy
-
-    def match_strings(self, expected: str, output: str) -> bool:
-        if not self._case_sensitive:
-            expected = expected.lower()
-            output = output.lower()
-
-        if self._fuzzy:
-            return expected in output or output in expected
-
-        return expected == output
-
-    def evaluate_prediction(self, prediction: Prediction) -> Optional[Evaluator.Match]:
-        output = prediction.output
-        for expected in prediction.test.expected:
-            if self.match_strings(expected, output):
-                return Evaluator.Match(prediction=prediction.output, expected=expected)
-        return None
-
-
-def load_prediction_files(paths: List[Path]) -> List[Prediction]:
-    import json
-
-    import yaml
-
-    predictions = []
-    for path in paths:
-        for file_path in path.rglob("*"):
-            if not file_path.is_file():
-                continue
-            if file_path.suffix not in {".json", ".yml", ".yaml"}:
-                continue
-            with open(file_path, "r") as file:
-                if file_path.suffix == ".json":
-                    data = json.load(file)
-                    predictions.append(Prediction(**data))
-                elif file_path.suffix in {".yml", ".yaml"}:
-                    data = yaml.safe_load(file)
-                    predictions.append(Prediction(**data))
-    return predictions
